Improve the structure of the Test project * Same folder structure as Funcky * Split the EnumerableExtensionTests
