+++ Funcky.Test/Functional/CurryTest.cs
