+++ Funcky.Test/Functional/FlipTest.cs
