+++ Funcky.Test/Functional/NoOperationTest.cs
