+++ Funcky.Test/Functional/PredicateCompositionTest.cs
