+++ Funcky.Test/Functional/UncurryTest.cs
