+++ Funcky.Test/Monads/EitherTest.cs
