+++ Funcky.Test/Monads/ResultTest.cs
