+++ Funcky.Test/TestUtils/SideEffect.cs
